pathfinding off of main thread.
Also changed pathfinding to use indexes so that it would be easier to serialize, but I don't think it needs to be serialized anymore?

Currently building the octree from renderer bounds, which was pretty easy to add. If I need to go back to building from meshes I will have to solve the serialization. I may want to revert some of these changes but I don't think they hurt anything right now...